+++ modules/post/components/timeRange/type.ts
@@ -1,3 +1,4 @@
+import { MeetingDateRange } from '@post/payload/postPayload';
 import { Dispatch, SetStateAction } from 'react';
 
 export type PayloadType = 'creator' | 'modifier';
@@ -8,4 +9,5 @@ export interface ItimeRange {
   description: string;
   onPressDay: (setRange: Dispatch<SetStateAction<string | null>>) => void;
   payloadType: PayloadType;
+  prevRange?: MeetingDateRange;
 }
